--- a/docs/release-notes/release-notes-0.14.0.md
+++ b/docs/release-notes/release-notes-0.14.0.md
@@ -376,6 +376,9 @@ you.
 * [Cache the channel state](https://github.com/lightningnetwork/lnd/pull/5595)
   to achieve better performance when running LND using a remote DB backend.
 
+* [Do not re-fetch payments if we already have them in memory](https://github.com/lightningnetwork/lnd/pull/5769)
+  in certain cases.
+
 ## Log system
 
 * [Save compressed log files from logrorate during