**Edit:**
~- rebased on https://github.com/lightningnetwork/lnd/pull/5516 (1st commit)~
~- rebased on https://github.com/lightningnetwork/lnd/pull/5515 (2nd commit)~
~- rebased on https://github.com/lightningnetwork/lnd/pull/5514 (3rd commit)~
~- rebased on https://github.com/lightningnetwork/lnd/pull/5513 (4-5-6th commits)~

The results below are somewhat outdated since now all data is in the remote db.

- rebased on: https://github.com/lightningnetwork/lnd/pull/5595
- rebased on: https://github.com/lightningnetwork/lnd/pull/5769
~- rebased on: https://github.com/lightningnetwork/lnd/pull/5640~ (merged)

**Description:**
This PR adds a few performance improvements to LND's etcd `kvdb` wrapper and fixes a few of the most important hotspots.
The performance improvements is pretty dramatic and could possibly be further improved. Some of the changes require migration which is not part of the PR until ACKed.

Test 1: `make itest etcd=1 icase=async_payments_benchmark`

**master**
```
=== RUN   TestLightningNetworkDaemon/36-of-79/btcd/async_payments_benchmark
    test_harness.go:120: 	Benchmark info: Elapsed time:  5.851402397s
    test_harness.go:120: 	Benchmark info: TPS:  82.54431454716445
```

**bhandras:etcd_improvements**
```
=== RUN   TestLightningNetworkDaemon/36-of-79/btcd/async_payments_benchmark
    test_harness.go:120: 	Benchmark info: Elapsed time:  3.357472064s
    test_harness.go:120: 	Benchmark info: TPS:  143.85823345453753
```

For reference the same tests with bbolt:
**master**
```
=== RUN   TestLightningNetworkDaemon/36-of-79/btcd/async_payments_benchmark
    test_harness.go:120: 	Benchmark info: Elapsed time:  3.255560917s
    test_harness.go:120: 	Benchmark info: TPS:  148.3615304133472
```
**bhandras:etcd_improvements**
```
 === RUN   TestLightningNetworkDaemon/36-of-79/btcd/async_payments_benchmark
    test_harness.go:120: 	Benchmark info: Elapsed time:  2.841959831s
    test_harness.go:120: 	Benchmark info: TPS:  169.95314104423736
```

Test 2:
Bottlepay benchmark (https://github.com/bottlepay/lightning-benchmark/pull/6) `lnd-etcd-notls` with etcd 3.5. (machine is a Hetzner AX41-nvme dedicated server):

```
loadtest_1    | 2021-06-25T20:28:24.266Z	INFO	Speed	{"tps": 64.53423786816828, "count": 2000, "avg_latency_sec": 1.551165605579}
loadtest_1    | 2021-06-25T20:28:39.926Z	INFO	Speed	{"tps": 63.8589896810007, "count": 3000, "avg_latency_sec": 1.563214304955}
loadtest_1    | 2021-06-25T20:28:55.677Z	INFO	Speed	{"tps": 63.48883943608307, "count": 4000, "avg_latency_sec": 1.5691106821979999}
loadtest_1    | 2021-06-25T20:29:11.351Z	INFO	Speed	{"tps": 63.799733551016345, "count": 5000, "avg_latency_sec": 1.575523050719}
loadtest_1    | 2021-06-25T20:29:33.243Z	INFO	Speed	{"tps": 45.6792330787301, "count": 6000, "avg_latency_sec": 2.125946590737}
loadtest_1    | 2021-06-25T20:30:00.740Z	INFO	Speed	{"tps": 36.36786498886787, "count": 7000, "avg_latency_sec": 2.7182521204489998}
loadtest_1    | 2021-06-25T20:30:37.995Z	INFO	Speed	{"tps": 26.841979764258515, "count": 8000, "avg_latency_sec": 3.712007323702}
loadtest_1    | 2021-06-25T20:31:15.672Z	INFO	Speed	{"tps": 26.541241455872445, "count": 9000, "avg_latency_sec": 3.7649197021939997}
loadtest_1    | 2021-06-25T20:31:53.198Z	INFO	Speed	{"tps": 26.648147887337295, "count": 10000, "avg_latency_sec": 3.760037243958}
loadtest_1    | 2021-06-25T20:32:30.899Z	INFO	Speed	{"tps": 26.52495979994532, "count": 11000, "avg_latency_sec": 3.762526293845}
loadtest_1    | 2021-06-25T20:33:08.355Z	INFO	Speed	{"tps": 26.697739686752644, "count": 12000, "avg_latency_sec": 3.743134903228}
loadtest_1    | 2021-06-25T20:33:46.253Z	INFO	Speed	{"tps": 26.387032710041233, "count": 13000, "avg_latency_sec": 3.778454786527}
loadtest_1    | 2021-06-25T20:34:22.728Z	INFO	Speed	{"tps": 27.41602735140185, "count": 14000, "avg_latency_sec": 3.648309565185}
loadtest_1    | 2021-06-25T20:34:59.736Z	INFO	Speed	{"tps": 27.020954368414703, "count": 15000, "avg_latency_sec": 3.7169224409929997}
loadtest_1    | 2021-06-25T20:35:36.832Z	INFO	Speed	{"tps": 26.95708458772625, "count": 16000, "avg_latency_sec": 3.718038788403}
loadtest_1    | 2021-06-25T20:36:14.057Z	INFO	Speed	{"tps": 26.863801955015347, "count": 17000, "avg_latency_sec": 3.7005437773809997}
loadtest_1    | 2021-06-25T20:36:50.769Z	INFO	Speed	{"tps": 27.23912651754751, "count": 18000, "avg_latency_sec": 3.702418314502}
loadtest_1    | 2021-06-25T20:37:27.028Z	INFO	Speed	{"tps": 27.579645991685165, "count": 19000, "avg_latency_sec": 3.62833937356}
loadtest_1    | 2021-06-25T20:38:03.640Z	INFO	Speed	{"tps": 27.31288613225411, "count": 20000, "avg_latency_sec": 3.640910035547}
loadtest_1    | 2021-06-25T20:38:38.998Z	INFO	Speed	{"tps": 28.28272378039225, "count": 21000, "avg_latency_sec": 3.549461816367}
loadtest_1    | 2021-06-25T20:39:15.517Z	INFO	Speed	{"tps": 27.38268223134218, "count": 22000, "avg_latency_sec": 3.622988230753}
loadtest_1    | 2021-06-25T20:39:51.170Z	INFO	Speed	{"tps": 28.04850209294464, "count": 23000, "avg_latency_sec": 3.616076497207}
loadtest_1    | 2021-06-25T20:40:14.835Z	INFO	Speed	{"tps": 42.25609409199764, "count": 24000, "avg_latency_sec": 2.3856274190680002}
loadtest_1    | 2021-06-25T20:40:41.624Z	INFO	Speed	{"tps": 37.328661146707326, "count": 25000, "avg_latency_sec": 2.631493367686}
loadtest_1    | 2021-06-25T20:41:16.835Z	INFO	Speed	{"tps": 28.40056707352067, "count": 26000, "avg_latency_sec": 3.518374261111}
loadtest_1    | 2021-06-25T20:41:51.929Z	INFO	Speed	{"tps": 28.495038611114413, "count": 27000, "avg_latency_sec": 3.449567716871}
loadtest_1    | 2021-06-25T20:42:25.699Z	INFO	Speed	{"tps": 29.612065864047747, "count": 28000, "avg_latency_sec": 3.412843902324}
loadtest_1    | 2021-06-25T20:42:59.842Z	INFO	Speed	{"tps": 29.288941299676388, "count": 29000, "avg_latency_sec": 3.440998649995}
loadtest_1    | 2021-06-25T20:43:30.855Z	INFO	Speed	{"tps": 32.24437955314579, "count": 30000, "avg_latency_sec": 3.1074791252639997}
loadtest_1    | 2021-06-25T20:44:01.289Z	INFO	Speed	{"tps": 32.85841298524933, "count": 31000, "avg_latency_sec": 3.048717978335}
loadtest_1    | 2021-06-25T20:44:26.317Z	INFO	Speed	{"tps": 39.95553007358418, "count": 32000, "avg_latency_sec": 2.535706471033}
loadtest_1    | 2021-06-25T20:44:55.086Z	INFO	Speed	{"tps": 34.75919827405728, "count": 33000, "avg_latency_sec": 2.901374373912}
loadtest_1    | 2021-06-25T20:45:20.299Z	INFO	Speed	{"tps": 39.662256248877995, "count": 34000, "avg_latency_sec": 2.402428623677}
loadtest_1    | 2021-06-25T20:45:47.715Z	INFO	Speed	{"tps": 36.475168809779504, "count": 35000, "avg_latency_sec": 2.6934887744009997}
loadtest_1    | 2021-06-25T20:46:13.449Z	INFO	Speed	{"tps": 38.860213406542506, "count": 36000, "avg_latency_sec": 2.732327640684}
loadtest_1    | 2021-06-25T20:46:40.743Z	INFO	Speed	{"tps": 36.63767627400598, "count": 37000, "avg_latency_sec": 2.71133185894}
loadtest_1    | 2021-06-25T20:47:14.283Z	INFO	Speed	{"tps": 29.815589101065907, "count": 38000, "avg_latency_sec": 3.251210235977}
loadtest_1    | 2021-06-25T20:47:50.904Z	INFO	Speed	{"tps": 27.306501429238995, "count": 39000, "avg_latency_sec": 3.5618042077629997}
loadtest_1    | 2021-06-25T20:48:20.768Z	INFO	Speed	{"tps": 33.48506530134645, "count": 40000, "avg_latency_sec": 3.178217333703}
loadtest_1    | 2021-06-25T20:48:53.882Z	INFO	Speed	{"tps": 30.198867511743984, "count": 41000, "avg_latency_sec": 3.261326943087}
loadtest_1    | 2021-06-25T20:49:30.377Z	INFO	Speed	{"tps": 27.4009239644335, "count": 42000, "avg_latency_sec": 3.6618183438000003}
loadtest_1    | 2021-06-25T20:50:07.803Z	INFO	Speed	{"tps": 26.71946799992048, "count": 43000, "avg_latency_sec": 3.788799161978}
loadtest_1    | 2021-06-25T20:50:40.953Z	INFO	Speed	{"tps": 30.166464100137283, "count": 44000, "avg_latency_sec": 3.261829291491}
loadtest_1    | 2021-06-25T20:51:18.100Z	INFO	Speed	{"tps": 26.920071596310144, "count": 45000, "avg_latency_sec": 3.6476010820380003}
loadtest_1    | 2021-06-25T20:51:57.696Z	INFO	Speed	{"tps": 25.2552097384627, "count": 46000, "avg_latency_sec": 3.989441471565}
loadtest_1    | 2021-06-25T20:52:31.398Z	INFO	Speed	{"tps": 29.672012604051638, "count": 47000, "avg_latency_sec": 3.4311492841969997}
loadtest_1    | 2021-06-25T20:53:09.045Z	INFO	Speed	{"tps": 26.562714943040568, "count": 48000, "avg_latency_sec": 3.708941020582}
loadtest_1    | 2021-06-25T20:53:41.258Z	INFO	Speed	{"tps": 31.04299467250739, "count": 49000, "avg_latency_sec": 3.223344582401}
loadtest_1    | 2021-06-25T20:54:16.625Z	INFO	Speed	{"tps": 28.275201341899614, "count": 50000, "avg_latency_sec": 3.4959532821379997}
```

It's possible that there's still some amount of degradation over time. After better caching and prefetches at hotspots the main culprit seemed to be the payment sequence and the misson control store. In my tests temporally removing those components completely eliminated all degradation (fluctuations still remained).