@bhandras 2021-06-25
An idea to fix the mission control store could be to completely eliminate the use of the cursor, but instead extend the `kvdb` interface with a `DeleteRange(prefix, limit)` function. The current fix (last commit) is just a quick hack.