@joostjager 2021-07-09
I ran this branch (commit 12e2fc025d8d100df9238fa7dd2a0ced32ecbddc) again on the benchmark rig. Still seeing performance going down:

```
loadtest_1    | 2021-07-08T14:07:19.311Z        INFO    Speed   {"tps": 75.14193036238315, "count": 10000, "avg_latency_sec": 1.305311964003}
loadtest_1    | 2021-07-08T14:07:33.757Z        INFO    Speed   {"tps": 69.22095510665038, "count": 11000, "avg_latency_sec": 1.428266781691}
loadtest_1    | 2021-07-08T14:07:48.706Z        INFO    Speed   {"tps": 66.89598692150489, "count": 12000, "avg_latency_sec": 1.518786493139}
loadtest_1    | 2021-07-08T14:08:07.099Z        INFO    Speed   {"tps": 54.36768485892386, "count": 13000, "avg_latency_sec": 1.811761789857}
...
loadtest_1    | 2021-07-08T14:13:17.195Z        INFO    Speed   {"tps": 33.471198411439225, "count": 26000, "avg_latency_sec": 3.036198552882}
loadtest_1    | 2021-07-08T14:13:48.885Z        INFO    Speed   {"tps": 31.55597571425602, "count": 27000, "avg_latency_sec": 3.10192132433}
loadtest_1    | 2021-07-08T14:14:23.377Z        INFO    Speed   {"tps": 28.992107302965735, "count": 28000, "avg_latency_sec": 3.520457434735}
loadtest_1    | 2021-07-08T14:17:22.060Z        INFO    Speed   {"tps": 25.572979865416883, "count": 33000, "avg_latency_sec": 3.9895093127959997}
...
loadtest_1    | 2021-07-08T14:29:02.549Z        INFO    Speed   {"tps": 19.782371680872405, "count": 48000, "avg_latency_sec": 5.0474531862909995}
loadtest_1    | 2021-07-08T14:29:58.412Z        INFO    Speed   {"tps": 17.90082123591677, "count": 49000, "avg_latency_sec": 5.68911388522}
loadtest_1    | 2021-07-08T14:30:54.896Z        INFO    Speed   {"tps": 17.704329283512795, "count": 50000, "avg_latency_sec": 5.639053898379999}
loadtest_1    | 2021-07-08T14:31:53.200Z        INFO    Speed   {"tps": 17.15164829144177, "count": 51000, "avg_latency_sec": 5.727623284748}
...
loadtest_1    | 2021-07-08T15:00:16.638Z        INFO    Speed   {"tps": 10.80652597372745, "count": 74000, "avg_latency_sec": 8.320684109424}
loadtest_1    | 2021-07-08T15:01:42.287Z        INFO    Speed   {"tps": 11.675464192683888, "count": 75000, "avg_latency_sec": 9.581611573754}
loadtest_1    | 2021-07-08T15:03:19.586Z        INFO    Speed   {"tps": 10.277611002407355, "count": 76000, "avg_latency_sec": 8.670611182626}
loadtest_1    | 2021-07-08T15:04:44.056Z        INFO    Speed   {"tps": 11.838525160308704, "count": 77000, "avg_latency_sec": 9.304994055154001}
```

The benchmark eventually crashes. Some of the errors in the log:

`[ERR] HSWC: Unhandled error while reforwarding htlc settle/fail over htlcswitch: circuit has already been closed`

`[ERR] CRTR: Failed to update mission control store: etcd error: stm.Commit() failed - etcdserver: request timed out`

`[WRN] HSWC: ChannelLink(508:8:0): unable to remove fwd pkgs: etcd error: stm.Commit() failed - etcdserver: too many operations in txn request`