@bhandras 2021-07-09
You need to compact and defrag since with that many payments we'll keep a lot of history in etcd. Or you need to set autocompact and defrag.