@bhandras 2021-07-09
Also interesting that I'm not able to reproduce the same degradation as you (ptal included results above). What version of etcd did you use? Is this the same gcloud instance?