@joostjager 2021-07-09
>You need to compact and defrag since with that many payments we'll keep a lot of history in etcd. Or you need to set autocompact and defrag.

How to configure this exactly?

>Also interesting that I'm not able to reproduce the same degradation as you (ptal included results above). What version of etcd did you use? Is this the same gcloud instance?

I use the code that you supplied to me earlier: https://github.com/bottlepay/lightning-benchmark/compare/etcd-improvements-test. Running on the same gcloud instance.

Maybe you can give me a patch on top of that so that I am surely running the right config?